small style fixes

+++ module1/solutions/task1.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-size_t multiply(int lhs, int rhs) {
+int multiply(int lhs, int rhs) {
     return lhs * rhs;
 }
